Fixed week 4 folder
I looked at the changes y'all made, and Week 5 had all of the files, but Week 4 only had tetra.obj in it's folder. So, I made a copy of the Week 4 'Homework_Everyone' folder and fixed the Week 4 folder. I also added the PDF file from Professor Martin into the folder for Week 4 and Week 5 :)